Testing optional TTL functionality without interface
